♻️ refactor logger
